ec2_lgw_route_table_association: add new module for Outposts
